select nodes by clicking on map polygons
had to split getVisibleNodes and sortVisibleNodesByColumn to have a
flat version of visibleNodes stored in the app store.
